fix(mojo-0263): replace capturing closures with NumericalForward structs
Mojo 0.26.3 disallows passing non-escaping (capturing/unified) closures
to `def(AnyTensor) raises -> AnyTensor` function parameters. Resolves 4
compile errors in layer_testers.mojo by introducing the NumericalForward
trait in gradient_checker.mojo and wrapping each capturing closure in a
struct that implements the trait.

Also fixes:
- stdlib imports: `from testing import` → `from std.testing import` etc.
- `raises escaping` → `raises` (escaping keyword removed in 0.26.3)
- `unified {read}` annotations on capturing nested defs in test files
- Unused loop variable `i` → `_` in test_memory_leaks.mojo
- Implicit Int→UInt conversions in test_unsigned.mojo
- Trailing whitespace

+++ tests/shared/core/test_unsigned.mojo
@@ -277,10 +277,10 @@ def test_uint_to_int_conversion() raises:
 def test_uint_from_int_conversion() raises:
     """Test constructing unsigned integers from Int values."""
     var i: Int = 42
-    var u8: UInt8 = i
-    var u16: UInt16 = i
-    var u32: UInt32 = i
-    var u64: UInt64 = i
+    var u8: UInt8 = UInt8(i)
+    var u16: UInt16 = UInt16(i)
+    var u32: UInt32 = UInt32(i)
+    var u64: UInt64 = UInt64(i)
 
     if u8 != 42:
         raise Error("Int -> UInt8 conversion failed")
